requirements.txt
flask
ultralytics
google-generativeai
python-dotenv